Fixed strange collisions with blocks.
Apparently this was happening because the player's hitbox components
were not added to the state. The old version of playstate probably handled
it in the past, so things broke when those lines were removed.

This fixes the issue by making the player responsible for adding their hitboxes
to the state themselves.